+++ android/gradle.properties
@@ -1,3 +1,4 @@
 org.gradle.jvmargs=-Xmx1536M
+android.enableR8=true
 android.useAndroidX=true
 android.enableJetifier=true
