--- lib/data/data_sub_tingkat_2.dart
@@ -1,538 +1,0 @@
-part of 'data.dart';
-
-const String dapow = 'Dasar Power';
-const String dager = 'Dasar Getaran';
-
-final _pKuda = SubTingkat2(
-  name: 'Kuda-kuda',
-  list: ['Hadap'],
-);
-final _pSiPas = SubTingkat2(
-  name: 'Sikap Pasang',
-  list: ['I Sampai VI', 'XII'],
-);
-final _pTangkis = SubTingkat2(
-  name: 'Tangkisan',
-  list: ['Buang Luar', 'Buang Dalam'],
-);
-final _pPukul = SubTingkat2(
-  name: 'Pukulan',
-  list: ['Lurus', 'Buku-buku', 'Tusuk Jari'],
-);
-final _pTendang = SubTingkat2(
-  name: 'Tendangan',
-  list: ['Lurus', 'Sabit', 'T', 'Belakang'],
-);
-final _pElak = SubTingkat2(
-  name: 'Elakan',
-  list: ['Ales'],
-);
-
-final _hitamTekDasKuda = SubTingkat2(
-  name: 'Kuda-kuda',
-  list: ['Serong', 'Pipih'],
-);
-
-final _hitamTekDasSiPas = SubTingkat2(
-  name: 'Sikap Pasang',
-  list: ['I Sampai XII'],
-);
-
-final _hitamTekDasTang = SubTingkat2(
-  name: 'Tangkisan',
-  list: ['Buang Luar dan Dalam', 'Jepit Atas dan Bawah', 'Kipat'],
-);
-
-final _hitamTekDasPuk = SubTingkat2(
-  name: 'Pukulan',
-  list: ['Lurus', 'Buku-buku', 'Tusuk Jari', 'Rangkap/Papas'],
-);
-
-final _hitamTekDasTen = SubTingkat2(
-  name: 'Tendangan',
-  list: ['Lurus', 'T', 'Sabit', 'Belakang', 'Serong'],
-);
-
-final _hitamTekDasElak = SubTingkat2(
-  name: 'Elakan',
-  list: ['Ales', 'Harimau'],
-);
-
-final _hitamPuSam = SubTingkat2(
-  name: 'Pukul-Sambut',
-  list: ['I Sampai IV'],
-);
-
-final _hitamSenJu = SubTingkat2(
-  name: 'Senam Jurus',
-  list: ['I Sampai VI'],
-);
-
-final _hitamDaPow = SubTingkat2(
-  name: '$dapow (Sasaran 1 Hebel)',
-  list: [],
-);
-
-final _hitamDaGer = SubTingkat2(
-  name: '$dager (Halang Rintang)',
-  list: [],
-);
-
-final _bKuda = SubTingkat2(
-  name: 'Kuda-kuda',
-  list: ['Pipih'],
-);
-
-final _bSiPas = SubTingkat2(
-  name: 'Sikap Pasang',
-  list: ['I Sampai XII'],
-);
-
-final _bLangkah = SubTingkat2(
-  name: 'Langkah',
-  list: ['Segitiga'],
-);
-
-final _bTangkis = SubTingkat2(
-  name: 'Tangkisan',
-  list: ['Kipat', 'Kepuk'],
-);
-
-final _bPukul = SubTingkat2(
-  name: 'Pukulan',
-  list: ['Lurus', 'Buku-buku', 'Tusukan Jari', 'Punggung Tangan'],
-);
-
-final _bTendang = SubTingkat2(
-  name: 'Tendangan',
-  list: ['Lurus', 'T', 'Sabit Lompat', 'Gejos'],
-);
-
-final _bElak = SubTingkat2(
-  name: 'Elakan',
-  list: ['Egos'],
-);
-
-final _bSikuan = SubTingkat2(
-  name: 'Sikuan',
-  list: ['Tusuk', 'Sangga'],
-);
-
-final _bRedam = SubTingkat2(
-  name: 'Redaman/Jatuhan',
-  list: ['Samping', 'Depan'],
-);
-
-final _bPuSam = SubTingkat2(
-  name: 'Pukul-Sambut',
-  list: ['V Sampai VIII'],
-);
-
-final _bSenJu = SubTingkat2(
-  name: 'Senam Jurus',
-  list: ['VII Sampai XII'],
-);
-
-final _bDaPow = SubTingkat2(
-  name: '$dapow (Sasaran 2 Hebel)',
-  list: [],
-);
-
-final _bDaGer = SubTingkat2(
-  name: '$dager (Halang Rintang)',
-  list: [],
-);
-
-final _b1SiPas = SubTingkat2(
-  name: 'Sikap Pasang',
-  list: ['VII Sampai XII'],
-);
-
-final _b1Langkah = SubTingkat2(
-  name: 'Langkah',
-  list: ['Zig-zag'],
-);
-
-final _b1Tangkis = SubTingkat2(
-  name: 'Tangkisan',
-  list: ['Sangga', 'Kepruk Gantung'],
-);
-
-final _b1Pukul = SubTingkat2(
-  name: 'Pukulan',
-  list: ['Sanggah', 'Papas'],
-);
-
-final _b1Tendang = SubTingkat2(
-  name: 'Tendangan',
-  list: ['Serong', 'Gejos'],
-);
-
-final _b1Elak = SubTingkat2(
-  name: 'Elakan',
-  list: ['Egos'],
-);
-
-final _b1Sikuan = SubTingkat2(
-  name: 'Sikuan',
-  list: ['Samping Dalam', 'Samping Luar'],
-);
-
-final _b1Redam = SubTingkat2(
-  name: 'Redaman/Jatuhan',
-  list: ['Guling Depan', 'Belakang'],
-);
-
-final _b1TangKa = SubTingkat2(
-  name: 'Tangkapan Kaki',
-  list: ['Dalam'],
-);
-
-final _b1PuSam = SubTingkat2(
-  name: 'Pukul-Sambut',
-  list: ['VII Sampai IX'],
-);
-
-final _b1JuPer = SubTingkat2(
-  name: 'Jurus Pertandingan',
-  list: ['I Sampai IV'],
-);
-
-final _b1Tuntun = SubTingkat2(
-  name: 'Tuntun',
-  list: ['1 Serangan'],
-);
-
-final _b1Solo = SubTingkat2(
-  name: 'Jurus Tunggal/Solo',
-  list: ['I (Satu)'],
-);
-
-final _b1DaPow = SubTingkat2(
-  name: '$dapow (Sasaran 2 Hebel, 1 Besi Dragon R-1)',
-  list: [],
-);
-
-final _b1DaGer = SubTingkat2(
-  name: '$dager (Halang Rintang)',
-  list: [],
-);
-
-final _b2Kuda = SubTingkat2(
-  name: 'Kuda-kuda',
-  list: ['Belakang'],
-);
-
-final _b2Langkah = SubTingkat2(
-  name: 'Langkah',
-  list: ['Ladam (U)'],
-);
-
-final _b2Tangkis = SubTingkat2(
-  name: 'Tangkisan',
-  list: ['Potong'],
-);
-
-final _b2Pukul = SubTingkat2(
-  name: 'Pukulan',
-  list: ['Ayun', 'Sudut'],
-);
-
-final _b2Tendang = SubTingkat2(
-  name: 'Tendangan',
-  list: ['Putar'],
-);
-
-final _b2Elak = SubTingkat2(
-  name: 'Elakan',
-  list: ['Selulup (Menyelam)'],
-);
-
-final _b2Sikuan = SubTingkat2(
-  name: 'Sikuan',
-  list: ['Bawah', 'Belakang'],
-);
-
-final _b2Redam = SubTingkat2(
-  name: 'Redaman',
-  list: ['Guling Belakang', 'Punggung'],
-);
-
-final _b2TangKa = SubTingkat2(
-  name: 'Tangkapan Kaki',
-  list: ['Luar'],
-);
-
-final _b2JuPer = SubTingkat2(
-  name: 'Jurus Pertandingan',
-  list: ['V Sampai IX'],
-);
-
-final _b2Tuntun = SubTingkat2(
-  name: 'Tuntun',
-  list: ['2 Serangan'],
-);
-
-final _samPer = SubTingkat2(
-  name: 'Sambung Pertandingan Versi IPSI',
-  list: [],
-);
-
-final _b2DaPow = SubTingkat2(
-  name: '$dapow (Sasaran 2 Hebel, 2 Besi Dragon R-1)',
-  list: [],
-);
-
-final _b2DaGer = SubTingkat2(
-  name: '$dager (Halang Rintang)',
-  list: [],
-);
-
-final _hKuda = SubTingkat2(
-  name: 'Kuda-kuda',
-  list: ['Betawi'],
-);
-
-final _hLangkah = SubTingkat2(
-  name: 'Langkah',
-  list: ['Curi'],
-);
-
-final _hTangkis = SubTingkat2(
-  name: 'Tangkisan',
-  list: ['Liuk'],
-);
-
-final _hPukul = SubTingkat2(
-  name: 'Pukulan',
-  list: ['Gebang Dalam', 'Gebang Luar'],
-);
-
-final _hTendang = SubTingkat2(
-  name: 'Tendangan',
-  list: [
-    'Teratai',
-    'Belakang',
-    'T Bawah',
-  ],
-);
-
-final _hElak = SubTingkat2(
-  name: 'Elakan',
-  list: ['Liuk'],
-);
-
-final _hSikuan = SubTingkat2(
-  name: 'Sikuan',
-  list: ['Belah'],
-);
-
-final _hJatuh = SubTingkat2(
-  name: 'Jatuhan/Bantingan',
-  list: ['Sapuan Tegak', 'Koset'],
-);
-
-final _hKunci = SubTingkat2(
-  name: 'Kuncian',
-  list: ['Pergelangan Tangan'],
-);
-
-final _hRedam = SubTingkat2(
-  name: 'Redaman',
-  list: ['Meroda'],
-);
-
-final _hJuPer = SubTingkat2(
-  name: 'Jurus Pertandingan',
-  list: ['X Sampai XIV'],
-);
-
-final _hJuPem = SubTingkat2(
-  name: 'Jurus Pembelaan',
-  list: ['1 A,B,C'],
-);
-
-final _hOlah = SubTingkat2(
-  name: 'Olah Nafas',
-  list: ['Pengerasan Tubuh 1 Serangan'],
-);
-
-final _hSolo = SubTingkat2(
-  name: 'Jurus Tunggal/Solo',
-  list: ['2 (Dua)'],
-);
-
-final _hJuTung = SubTingkat2(
-  name: 'Jurus Tunggal IPSI',
-  list: [],
-);
-
-final _hDaPow = SubTingkat2(
-  name: '$dapow (2 Besi Dragon R-1, 1 Plat Baja K-1)',
-  list: [],
-);
-
-final _hDaGer = SubTingkat2(
-  name: '$dager (Halang Rintang)',
-  list: [],
-);
-
-final _h1Kuda = SubTingkat2(
-  name: 'Kuda-kuda',
-  list: ['Pilin'],
-);
-
-final _h1Langkah = SubTingkat2(
-  name: 'Langkah',
-  list: ['Kucing', 'Tipu'],
-);
-
-final _h1Tangkis = SubTingkat2(
-  name: 'Tangkisan',
-  list: ['Lenggang Bawah', 'Lenggang Atas', 'Kibas Tangan', 'Kibas Sisik'],
-);
-
-final _h1Pukul = SubTingkat2(
-  name: 'Pukulan',
-  list: ['Tebang Dalam', 'Tebang ke Luar', 'Tampar Dalam', 'Tampar ke Luar'],
-);
-
-final _h1Tendang = SubTingkat2(
-  name: 'Tendangan',
-  list: ['Pacul', 'Taji'],
-);
-
-final _h1Lutut = SubTingkat2(
-  name: 'Lututan',
-  list: ['Tusuk'],
-);
-
-final _h1Jatuh = SubTingkat2(
-  name: 'Jatuhan',
-  list: ['Sapuan Rebah Depan', 'Sapuan Rebah Belakang'],
-);
-
-final _h1Kunci = SubTingkat2(
-  name: 'Kuncian',
-  list: ['Siku'],
-);
-
-final _h1Redam = SubTingkat2(
-  name: 'Redaman',
-  list: ['Melenting'],
-);
-
-final _h1JuPer = SubTingkat2(
-  name: 'Jurus Pertandingan',
-  list: ['XV Sampai XIX'],
-);
-
-final _h1JuPem = SubTingkat2(
-  name: 'Jurus Pembelaan',
-  list: ['2 A,B,C'],
-);
-
-final _h1Olah = SubTingkat2(
-  name: 'Olah Nafas',
-  list: ['Pengerasan Tubuh 2 Serangan'],
-);
-
-final _h1KemBas = SubTingkat2(
-  name: 'Kembangan Bebas',
-  list: ['Minimal 10 Rangkaian'],
-);
-
-final _h1JuReg = SubTingkat2(
-  name: 'Jurus Regu Pesilat',
-  list: ['I Sampai VI'],
-);
-
-final _h1DaPow = SubTingkat2(
-  name: '$dapow (2 Besi Dragon R-1, 1 Besi Dragon R-2, 1 Plat Baja K-1)',
-  list: [],
-);
-
-final _h1DaGer = SubTingkat2(
-  name: '$dager (Halang Rintang)',
-  list: [],
-);
-
-final _h2Kuda = SubTingkat2(
-  name: 'Kuda-kuda',
-  list: ['Silang'],
-);
-
-final _h2Langkah = SubTingkat2(
-  name: 'Langkah',
-  list: ['Segiempat', 'Minang'],
-);
-
-final _h2Tangkis = SubTingkat2(
-  name: 'Tangkisan',
-  list: ['Caram', 'Kibas Kaki Dalam', 'Kibas Kaki Luar', 'Tepis'],
-);
-
-final _h2Pukul = SubTingkat2(
-  name: 'Pukulan',
-  list: ['Kepruk', 'Kedut', 'Palu', 'Belah'],
-);
-
-final _h2Tendang = SubTingkat2(
-  name: 'Tendangan',
-  list: ['Gaet'],
-);
-
-final _h2Lutut = SubTingkat2(
-  name: 'Lututan',
-  list: ['Putar'],
-);
-
-final _h2Jatuh = SubTingkat2(
-  name: 'Jatuhan',
-  list: ['Ungkitan', 'Kaitan', 'Sabetan'],
-);
-
-final _h2Kunci = SubTingkat2(
-  name: 'Kuncian',
-  list: ['Bahu'],
-);
-
-final _h2Redam = SubTingkat2(
-  name: 'Redaman',
-  list: ['Lompat Harimau'],
-);
-
-final _h2JuPer = SubTingkat2(
-  name: 'Jurus Pertandingan',
-  list: ['XX Sampai XXIV'],
-);
-
-final _h2JuPem = SubTingkat2(
-  name: 'Jurus Pembelaan',
-  list: ['3 A,B,C'],
-);
-
-final _h2Olah = SubTingkat2(
-  name: 'Olah Nafas',
-  list: ['Pengerasan Tubuh 2 Penyerang'],
-);
-
-final _h2KemBas = SubTingkat2(
-  name: 'Kembangan Bebas',
-  list: ['Minimal 15 Rangkaian'],
-);
-
-final _h2JuReg = SubTingkat2(
-  name: 'Jurus Regu Pesilat',
-  list: ['Lengkap'],
-);
-
-final _h2DaPow = SubTingkat2(
-  name: '$dapow (2 Besi Dragon R-1, 2 Besi Dragon R-2, 1 Plat Baja K-1)',
-  list: [],
-);
-
-final _h2DaGer = SubTingkat2(
-  name: '$dager (Halang Rintang)',
-  list: [],
-);
