+++ lib/data/materi/data_golongan.dart
@@ -1,0 +1,68 @@
+part of 'data.dart';
+
+final _calonSatria = Golongan(
+  imageUrl: 'assets/image/calon_satria.png',
+  name: 'Golongan Calon Satria',
+  tingkat: [
+    _pelantikan,
+    _sabukHitam,
+  ],
+);
+
+final _satriaMuda = Golongan(
+  imageUrl: 'assets/image/Kesatria (Icon 2-4).png',
+  name: 'Golongan Satria Muda',
+  tingkat: [
+    _sabukBiru,
+    _sabukBiru1,
+    _sabukBiru2,
+  ],
+);
+
+final _satriaMadya = Golongan(
+  imageUrl: 'assets/image/Kesatria (Icon 2-4).png',
+  name: 'Golongan Satria Madya',
+  tingkat: [
+    _sabukHijau,
+    _sabukHijau1,
+    _sabukHijau2,
+  ],
+);
+
+final _satriaUtama = Golongan(
+  imageUrl: 'assets/image/Kesatria (Icon 2-4).png',
+  name: 'Golongan Satria Utama',
+  tingkat: [
+    _sabukCokelat,
+    _sabukCokelat1,
+    _sabukCokelat2,
+  ],
+);
+
+final _pandegaMuda = Golongan(
+  imageUrl: 'assets/image/Pandega (Icon 5-7).png',
+  name: 'Golongan Pandega Muda',
+  tingkat: [
+    _sabukUngu,
+    _sabukUngu1,
+    _sabukUngu2,
+  ],
+);
+
+final _pandegaMadya = Golongan(
+  imageUrl: 'assets/image/Pandega (Icon 5-7).png',
+  name: 'Golongan Pandega Madya',
+  tingkat: [_sabukMerah],
+);
+
+final _pandegaUtama = Golongan(
+  imageUrl: 'assets/image/Pandega (Icon 5-7).png',
+  name: 'Golongan Pandega Utama',
+  tingkat: [_sabukJingga],
+);
+
+final _pendekarMuda = Golongan(
+  imageUrl: 'assets/image/Pendekar (Icon 8).png',
+  name: 'Golongan Pendekar Muda',
+  tingkat: [_sabukKuning],
+);
