+++ lib/data/materi/data_sub_tingkat.dart
@@ -1,0 +1,297 @@
+part of 'data.dart';
+
+final _pTeknikDasar = SubTingkat(
+  subTingkatList: [_pKuda, _pSiPas, _pTangkis, _pPukul, _pTendang, _pElak],
+);
+
+final _pLatihanBeladiri = SubTingkat(subTingkatList: []);
+final _pKeindahanGerak = SubTingkat(subTingkatList: []);
+final _pKetangkasan = SubTingkat(subTingkatList: []);
+final _pOlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _hitamTeknikDasar = SubTingkat(
+  subTingkatList: [
+    _hitamTekDasKuda,
+    _hitamTekDasSiPas,
+    _hitamTekDasTang,
+    _hitamTekDasPuk,
+    _hitamTekDasTen,
+    _hitamTekDasElak
+  ],
+);
+
+final _hitamLatihanBeladiri = SubTingkat(subTingkatList: [_hitamPuSam]);
+
+final _hitamKeindahanGerak = SubTingkat(subTingkatList: [_hitamSenJu]);
+
+final _hitamKetangkasan = SubTingkat(subTingkatList: []);
+
+final _hitamOlahPernafasan = SubTingkat(
+  subTingkatList: [
+    _hitamDaPow,
+    _hitamDaGer,
+  ],
+);
+
+final _bTeknikDasar = SubTingkat(
+  subTingkatList: [
+    _bKuda,
+    _bSiPas,
+    _bLangkah,
+    _bTangkis,
+    _bPukul,
+    _bTendang,
+    _bElak,
+    _bSikuan,
+    _bRedam,
+  ],
+);
+
+final _bLatihanBeladiri = SubTingkat(
+  subTingkatList: [_bPuSam],
+);
+
+final _bKeindahanGerak = SubTingkat(
+  subTingkatList: [_bSenJu],
+);
+
+final _bKetangkasan = SubTingkat(subTingkatList: []);
+
+final _bOlahPernafasan = SubTingkat(
+  subTingkatList: [
+    _bDaPow,
+    _bDaGer,
+  ],
+);
+
+final _b1TeknikDasar = SubTingkat(
+  subTingkatList: [
+    _b1SiPas,
+    _b1Langkah,
+    _b1Tangkis,
+    _b1Pukul,
+    _b1Tendang,
+    _b1Elak,
+    _b1Sikuan,
+    _b1Redam,
+    _b1TangKa,
+  ],
+);
+
+final _b1LatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    _b1PuSam,
+    _b1JuPer,
+    _b1Tuntun,
+  ],
+);
+
+final _b1KeindahanGerak = SubTingkat(subTingkatList: [_b1Solo]);
+
+final _b1Ketangkasan = SubTingkat(subTingkatList: [_samPer]);
+
+final _b1OlahPernafasan = SubTingkat(
+  subTingkatList: [
+    _b1DaPow,
+    _b1DaGer,
+  ],
+);
+
+final _b2TeknikDasar = SubTingkat(subTingkatList: [
+  _b2Kuda,
+  _b2Langkah,
+  _b2Tangkis,
+  _b2Pukul,
+  _b2Tendang,
+  _b2Elak,
+  _b2Sikuan,
+  _b2Redam,
+  _b2TangKa,
+]);
+
+final _b2LatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    _b2JuPer,
+    _b2Tuntun,
+  ],
+);
+
+final _b2KeindahanGerak = SubTingkat(subTingkatList: []);
+
+final _b2Ketangkasan = SubTingkat(subTingkatList: [_samPer]);
+
+final _b2OlahPernafasan = SubTingkat(
+  subTingkatList: [
+    _b2DaPow,
+    _b2DaGer,
+  ],
+);
+
+final _hTeknikDasar = SubTingkat(
+  subTingkatList: [
+    _hKuda,
+    _hLangkah,
+    _hTangkis,
+    _hPukul,
+    _hTendang,
+    _hElak,
+    _hSikuan,
+    _hJatuh,
+    _hKunci,
+    _hRedam,
+  ],
+);
+
+final _hLatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    _hJuPer,
+    _hJuPem,
+    _hOlah,
+  ],
+);
+
+final _hKeindahanGerak = SubTingkat(
+  subTingkatList: [
+    _hSolo,
+    _hJuTung,
+  ],
+);
+
+final _hKetangkasan = SubTingkat(subTingkatList: [_samPer]);
+
+final _hOlahPernafasan = SubTingkat(
+  subTingkatList: [
+    _hDaPow,
+    _hDaGer,
+  ],
+);
+
+final _h1TeknikDasar = SubTingkat(
+  subTingkatList: [
+    _h1Kuda,
+    _h1Langkah,
+    _h1Tangkis,
+    _h1Pukul,
+    _h1Tendang,
+    _h1Lutut,
+    _h1Jatuh,
+    _h1Kunci,
+    _h1Redam
+  ],
+);
+
+final _h1LatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    _h1JuPer,
+    _h1JuPem,
+    _h1Olah,
+  ],
+);
+
+final _h1KeindahanGerak = SubTingkat(
+  subTingkatList: [
+    _h1KemBas,
+    _h1JuReg,
+  ],
+);
+
+final _h1Ketangkasan = SubTingkat(subTingkatList: [_samPer]);
+
+final _h1OlahPernafasan = SubTingkat(
+  subTingkatList: [
+    _h1DaPow,
+    _h1DaGer,
+  ],
+);
+
+final _h2TeknikDasar = SubTingkat(
+  subTingkatList: [
+    _h2Kuda,
+    _h2Langkah,
+    _h2Tangkis,
+    _h2Pukul,
+    _h2Tendang,
+    _h2Lutut,
+    _h2Jatuh,
+    _h2Kunci,
+    _h2Redam,
+  ],
+);
+
+final _h2LatihanBeladiri = SubTingkat(
+  subTingkatList: [
+    _h2JuPer,
+    _h2JuPem,
+    _h2Olah,
+  ],
+);
+
+final _h2KeindahanGerak = SubTingkat(
+  subTingkatList: [
+    _h2KemBas,
+    _h2JuReg,
+  ],
+);
+
+final _h2Ketangkasan = SubTingkat(subTingkatList: [_samPer]);
+
+final _h2OlahPernafasan = SubTingkat(
+  subTingkatList: [
+    _h2DaPow,
+    _h2DaGer,
+  ],
+);
+
+final _cTeknikDasar = SubTingkat(subTingkatList: []);
+final _cLatihanBeladiri = SubTingkat(subTingkatList: []);
+final _cKeindahanGerak = SubTingkat(subTingkatList: []);
+final _cKetangkasan = SubTingkat(subTingkatList: []);
+final _cOlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _c1TeknikDasar = SubTingkat(subTingkatList: []);
+final _c1LatihanBeladiri = SubTingkat(subTingkatList: []);
+final _c1KeindahanGerak = SubTingkat(subTingkatList: []);
+final _c1Ketangkasan = SubTingkat(subTingkatList: []);
+final _c1OlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _c2TeknikDasar = SubTingkat(subTingkatList: []);
+final _c2LatihanBeladiri = SubTingkat(subTingkatList: []);
+final _c2KeindahanGerak = SubTingkat(subTingkatList: []);
+final _c2Ketangkasan = SubTingkat(subTingkatList: []);
+final _c2OlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _uTeknikDasar = SubTingkat(subTingkatList: []);
+final _uLatihanBeladiri = SubTingkat(subTingkatList: []);
+final _uKeindahanGerak = SubTingkat(subTingkatList: []);
+final _uKetangkasan = SubTingkat(subTingkatList: []);
+final _uOlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _u1TeknikDasar = SubTingkat(subTingkatList: []);
+final _u1LatihanBeladiri = SubTingkat(subTingkatList: []);
+final _u1KeindahanGerak = SubTingkat(subTingkatList: []);
+final _u1Ketangkasan = SubTingkat(subTingkatList: []);
+final _u1OlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _u2TeknikDasar = SubTingkat(subTingkatList: []);
+final _u2LatihanBeladiri = SubTingkat(subTingkatList: []);
+final _u2KeindahanGerak = SubTingkat(subTingkatList: []);
+final _u2Ketangkasan = SubTingkat(subTingkatList: []);
+final _u2OlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _mTeknikDasar = SubTingkat(subTingkatList: []);
+final _mLatihanBeladiri = SubTingkat(subTingkatList: []);
+final _mKeindahanGerak = SubTingkat(subTingkatList: []);
+final _mKetangkasan = SubTingkat(subTingkatList: []);
+final _mOlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _jTeknikDasar = SubTingkat(subTingkatList: []);
+final _jLatihanBeladiri = SubTingkat(subTingkatList: []);
+final _jKeindahanGerak = SubTingkat(subTingkatList: []);
+final _jKetangkasan = SubTingkat(subTingkatList: []);
+final _jOlahPernafasan = SubTingkat(subTingkatList: []);
+
+final _kTeknikDasar = SubTingkat(subTingkatList: []);
+final _kLatihanBeladiri = SubTingkat(subTingkatList: []);
+final _kKeindahanGerak = SubTingkat(subTingkatList: []);
+final _kKetangkasan = SubTingkat(subTingkatList: []);
+final _kOlahPernafasan = SubTingkat(subTingkatList: []);
