+++ lib/data/materi/data_tingkat.dart
@@ -1,0 +1,222 @@
+part of 'data.dart';
+
+final _pelantikan = Tingkat(
+  name: 'Pelantikan',
+  form: 'assets/image/FORM P.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _pTeknikDasar,
+    _pLatihanBeladiri,
+    _pKeindahanGerak,
+    _pKetangkasan,
+    _pOlahPernafasan,
+  ],
+);
+
+final _sabukHitam = Tingkat(
+  name: 'Sabuk Hitam',
+  form: 'assets/image/FORM HITAM.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _hitamTeknikDasar,
+    _hitamLatihanBeladiri,
+    _hitamKeindahanGerak,
+    _hitamKetangkasan,
+    _hitamOlahPernafasan
+  ],
+);
+
+final _sabukBiru = Tingkat(
+  name: 'Sabuk Biru',
+  form: 'assets/image/FORM B.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _bTeknikDasar,
+    _bLatihanBeladiri,
+    _bKeindahanGerak,
+    _bKetangkasan,
+    _bOlahPernafasan,
+  ],
+);
+
+final _sabukBiru1 = Tingkat(
+  name: 'Sabuk Biru Belati 1',
+  form: 'assets/image/FORM B1.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _b1TeknikDasar,
+    _b1LatihanBeladiri,
+    _b1KeindahanGerak,
+    _b1Ketangkasan,
+    _b1OlahPernafasan,
+  ],
+);
+
+final _sabukBiru2 = Tingkat(
+  name: 'Sabuk Biru Belati 2',
+  form: 'assets/image/FORM B2.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _b2TeknikDasar,
+    _b2LatihanBeladiri,
+    _b2KeindahanGerak,
+    _b2Ketangkasan,
+    _b2OlahPernafasan,
+  ],
+);
+
+final _sabukHijau = Tingkat(
+  name: 'Sabuk Hijau',
+  form: 'assets/image/FORM H.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _hTeknikDasar,
+    _hLatihanBeladiri,
+    _hKeindahanGerak,
+    _hKetangkasan,
+    _hOlahPernafasan,
+  ],
+);
+
+final _sabukHijau1 = Tingkat(
+  name: 'Sabuk Hijau Belati 1',
+  form: 'assets/image/FORM H1.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _h1TeknikDasar,
+    _h1LatihanBeladiri,
+    _h1KeindahanGerak,
+    _h1Ketangkasan,
+    _h1OlahPernafasan,
+  ],
+);
+
+final _sabukHijau2 = Tingkat(
+  name: 'Sabuk Hijau Belati 2',
+  form: 'assets/image/FORM H2.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _h2TeknikDasar,
+    _h2LatihanBeladiri,
+    _h2KeindahanGerak,
+    _h2Ketangkasan,
+    _h2OlahPernafasan,
+  ],
+);
+
+final _sabukCokelat = Tingkat(
+  name: 'Sabuk Cokelat',
+  form: 'assets/image/FORM C.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _cTeknikDasar,
+    _cLatihanBeladiri,
+    _cKeindahanGerak,
+    _cKetangkasan,
+    _cOlahPernafasan
+  ],
+);
+
+final _sabukCokelat1 = Tingkat(
+  name: 'Sabuk Cokelat Belati 1',
+  form: 'assets/image/FORM C1.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _c1TeknikDasar,
+    _c1LatihanBeladiri,
+    _c1KeindahanGerak,
+    _c1Ketangkasan,
+    _c1OlahPernafasan
+  ],
+);
+
+final _sabukCokelat2 = Tingkat(
+  name: 'Sabuk Cokelat Belati 2',
+  form: 'assets/image/FORM C2.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _c2TeknikDasar,
+    _c2LatihanBeladiri,
+    _c2KeindahanGerak,
+    _c2Ketangkasan,
+    _c2OlahPernafasan
+  ],
+);
+
+final _sabukUngu = Tingkat(
+  name: 'Sabuk Ungu',
+  form: 'assets/image/FORM U.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _uTeknikDasar,
+    _uLatihanBeladiri,
+    _uKeindahanGerak,
+    _uKetangkasan,
+    _uOlahPernafasan
+  ],
+);
+
+final _sabukUngu1 = Tingkat(
+  name: 'Sabuk Ungu Belati 1',
+  form: 'assets/image/FORM U1.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _u1TeknikDasar,
+    _u1LatihanBeladiri,
+    _u1KeindahanGerak,
+    _u1Ketangkasan,
+    _u1OlahPernafasan
+  ],
+);
+
+final _sabukUngu2 = Tingkat(
+  name: 'Sabuk Ungu Belati 2',
+  form: 'assets/image/FORM U2.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _u2TeknikDasar,
+    _u2LatihanBeladiri,
+    _u2KeindahanGerak,
+    _u2Ketangkasan,
+    _u2OlahPernafasan
+  ],
+);
+
+final _sabukMerah = Tingkat(
+  name: 'Sabuk Merah',
+  form: 'assets/image/FORM M.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _mTeknikDasar,
+    _mLatihanBeladiri,
+    _mKeindahanGerak,
+    _mKetangkasan,
+    _mOlahPernafasan
+  ],
+);
+
+final _sabukJingga = Tingkat(
+  name: 'Sabuk Jingga',
+  form: 'assets/image/FORM J.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _jTeknikDasar,
+    _jLatihanBeladiri,
+    _jKeindahanGerak,
+    _jKetangkasan,
+    _jOlahPernafasan
+  ],
+);
+
+final _sabukKuning = Tingkat(
+  name: 'Sabuk Kuning',
+  form: 'assets/image/FORM K.jpg',
+  header: _materiDetailHeaders,
+  subTingkat: [
+    _kTeknikDasar,
+    _kLatihanBeladiri,
+    _kKeindahanGerak,
+    _kKetangkasan,
+    _kOlahPernafasan
+  ],
+);
