+++ lib/data/materi/data_video.dart
@@ -1,0 +1,5 @@
+part of 'data.dart';
+
+const String _vidTest = 'assets/videos/test.mp4';
+const String _vidBjorka = 'assets/videos/bjorka.mp4';
+const String _vidBlueming = 'assets/videos/blueming.mp4';
