+++ lib/icons/icons.dart
@@ -1,0 +1,15 @@
+const iconTeknikDasar = 'assets/image/Icon Teknik Dasar.png';
+const iconLatihanBeladiri = 'assets/image/Icon Latihan Beladiri.png';
+const iconKeindahanGerak = 'assets/image/Icon Keindahan Gerak.png';
+const iconKetangkasan = 'assets/image/Icon Ketangkasan.png';
+const iconOlahPernafasan = 'assets/image/Icon Olah Pernafasan.png';
+const iconFormUjian = 'assets/image/Icon Form Ujian.png';
+
+final List<String> icons = [
+  iconTeknikDasar,
+  iconLatihanBeladiri,
+  iconKeindahanGerak,
+  iconKetangkasan,
+  iconOlahPernafasan,
+  iconFormUjian,
+];
