+++ lib/models/materi/sub_tingkat.dart
@@ -1,0 +1,9 @@
+part of 'golongan.dart';
+
+class SubTingkat {
+  final List<SubTingkat2> subTingkatList;
+
+  SubTingkat({
+    required this.subTingkatList,
+  });
+}
