+++ lib/models/materi/sub_tingkat_2.dart
@@ -1,0 +1,14 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
+part of 'golongan.dart';
+
+class SubTingkat2 {
+  final String name;
+  final List<String> list;
+  final List<String> vidAssets;
+
+  SubTingkat2({
+    required this.name,
+    required this.list,
+    required this.vidAssets,
+  });
+}
