--- lib/models/sub_tingkat.dart
@@ -1,9 +1,0 @@
-part 'sub_tingkat_2.dart';
-
-class SubTingkat {
-  final List<SubTingkat2> subTingkatList;
-
-  SubTingkat({
-    required this.subTingkatList,
-  });
-}
