--- lib/models/sub_tingkat_2.dart
@@ -1,12 +1,0 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
-part of 'sub_tingkat.dart';
-
-class SubTingkat2 {
-  final String name;
-  final List<String> list;
-
-  SubTingkat2({
-    required this.name,
-    required this.list,
-  });
-}
