+++ lib/screen/form_screen.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
 
-import '../models/tingkat.dart';
+import '../models/materi/golongan.dart';
 
 class FormScreen extends StatelessWidget {
   const FormScreen({Key? key, required this.tingkat}) : super(key: key);
