+++ pubspec.yaml
@@ -29,14 +29,16 @@ environment:
 # versions available, run `flutter pub outdated`.
 dependencies:
   animations: ^2.0.7
+  carousel_slider: ^4.2.1
+  chewie: ^1.4.0
   cupertino_icons: ^1.0.2
   dropdown_button2: ^1.9.2
+  flick_video_player: ^0.5.0
   fluentui_system_icons: ^1.1.190
   flutter:
     sdk: flutter
   google_fonts: ^4.0.3
   video_player: ^2.5.1
-  carousel_slider: ^4.2.1
 
 dev_dependencies:
   flutter_lints: ^2.0.0
