+++ test/widget_test.dart
@@ -13,7 +13,7 @@ import 'package:ksp_nusantara_app/main.dart';
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const Kesatria_PS_Nusantara());
+    await tester.pumpWidget(const KesatriaPSNusantara());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
